+++ CHANGELOG.md
@@ -12,6 +12,7 @@ For more information about changelogs, check
 * [FEATURE] Add `font_awesome_css` helper
 * [FEATURE] Add `dropdown` helper
 * [FEATURE] Add `progress_bar` helper
+* [ENHANCEMENT] Add `:fieldset` option to decide whether `fields_for` should wrap fields in a <fieldset> tag
 
 ## 1.0.1 - 2014-09-14