+++ CHANGELOG.md
@@ -6,7 +6,7 @@ For more information about changelogs, check
 [Keep a Changelog](http://keepachangelog.com) and
 [Vandamme](http://tech-angels.github.io/vandamme).
 
-## 1.1.0 - unreleased
+## 1.1.0 - 2014-09-20
 
 * [FEATURE] Add `icon` helper
 * [FEATURE] Add `font_awesome_css` helper